refactor(bin): extract file create and update into module
