+++ src/bin/commands/init.ts
@@ -1,13 +1,14 @@
 import fs from "node:fs/promises";
 import path from "node:path";
-import { fileURLToPath } from "node:url";
 import { Command } from "commander";
-import ejs from "ejs";
-import { pathExists } from "fs-extra";
 import ora from "ora";
-import { BIOME_CONFIG, TS_CONFIG } from "../constants";
-import { parseJson } from "../utils/json";
 import { initSchema } from "./schemas";
+import {
+  createBiomeConfig,
+  createTemplateFile,
+  createTsConfig,
+  updatePackageJson,
+} from "../files";
 
 // biome-ignore lint/suspicious/noExplicitAny: The type of options is not known at this point, so we use any.
 async function initAction(projectName: string, options: any) {
@@ -44,74 +45,6 @@ async function initAction(projectName: string, options: any) {
   }
 }
 
-async function updatePackageJson(projectRoot: string) {
-  const packageJsonPath = path.join(projectRoot, "package.json");
-  const doesPackageJsonExist = await pathExists(packageJsonPath);
-
-  if (!doesPackageJsonExist) {
-    console.warn("package.json does not exist, creating a new one.");
-  }
-
-  const existingPackageJsonContent = await fs.readFile(packageJsonPath, "utf8");
-  const existingPackageJson = JSON.parse(existingPackageJsonContent);
-
-  const newScripts = {
-    start: "node dist/index.mjs",
-    dev: "onlykit dev ./src",
-    build: "onlykit build ./src/index.ts",
-    check: "onlykit check",
-    "check:write": "onlykit check --write",
-  };
-  existingPackageJson.scripts = {
-    ...existingPackageJson.scripts,
-    ...newScripts,
-  };
-
-  await fs.writeFile(
-    path.join(projectRoot, "package.json"),
-    parseJson(existingPackageJson),
-    "utf8",
-  );
-}
-
-async function createTsConfig(projectRoot: string) {
-  await fs.writeFile(
-    path.join(projectRoot, "tsconfig.json"),
-    parseJson(TS_CONFIG),
-    "utf8",
-  );
-}
-
-async function createBiomeConfig(projectRoot: string) {
-  await fs.writeFile(
-    path.join(projectRoot, "biome.json"),
-    parseJson(BIOME_CONFIG),
-    "utf8",
-  );
-}
-
-async function createTemplateFile(
-  projectName: string,
-  srcDir: string,
-  template: string,
-) {
-  const __filename = fileURLToPath(import.meta.url);
-  const __dirname = path.dirname(__filename);
-  const templatePath = path.resolve(
-    __dirname,
-    "./templates",
-    template,
-    "index.ts.ejs",
-  );
-
-  const templateSource = await fs.readFile(templatePath, "utf8");
-  const rendered = ejs.render(templateSource, {
-    projectName,
-  });
-
-  await fs.writeFile(path.join(srcDir, "index.ts"), rendered, "utf8");
-}
-
 export const initCommand = new Command()
   .command("init")
   .description("Initialize a new OnlyKit project")
@@ -120,6 +53,6 @@ export const initCommand = new Command()
   .option(
     "-c, --cwd <path>",
     "Set the current working directory",
-    process.cwd(),
+    process.cwd()
   )
   .action(initAction);
