+++ src/bin/files/index.ts
@@ -1,0 +1,3 @@
+export * from "./config";
+export * from "./package";
+export * from "./template";
